Deployment: new script for etherbone&saftlib build and deployment
The script creates a build root and binaries using a specific
branch/commit of bel_projects. So far these scripts are though
for deployment of releases or all the timing stack together.
In a future, it will be possible to build and deploy the different
pieces of the stask (etherbone, drivers or saftlib)