build-rte: add git tag checkout
add possibility to checkout a tag on a branch,
if empty, it checks out the last commit of the
branch.

removes the "dirty/nothing to commit" tag for
the ci_cd repository. Anytime the script build_rte
is execute, it has been previously modified eg.
TARGET_DIR and it is going to be always dirty...