test-ctl: added new "metric" option to send TR metrics to Graphite/Grafana
Metrics obtained from the timing devices in the timing test facility are
sent periodically to Graphite using the UDP socket. Configuration file
for Graphite can be used to specify address, port and polling interval.
Graphite is backend database for the timing metrics and Grafa is used to
present the metrics graphically.