test-metric-ctl: added new script to send TR metrics to Graphite/Grafana
Metrics obtained from the timing devices in the timing test facility are
sent periodically to Graphite using the UDP socket. Configuration file
for Graphite can be used to specify address, port, polling interval and
how the WR sync status is mapped to numeric value.
Graphite is backend database for the timing metrics and Grafana is used to
present the metrics graphically.